Use addons endpoint to check for compute size instead
